extract: logic from SetUpDateStep
All the logic has been extracted into its context store.
The component has been divided into constituent subcomponents -
heading, snapshot picker, dependency notification.

The store has been placed above the template base modal.
